Fix missing type & tree annotators in InitializationFieldAccessChecker
